chore(build): move gulp tasks to their own directory
