+++ gulp/tasks/clean.js
@@ -1,0 +1,6 @@
+var gulp = require('gulp');
+var del = require('del');
+
+gulp.task('clean', function (done) {
+  del(['dist','demo/build'], done);
+});
